requirements.txt
fastapi
fastapi-login
SQLAlchemy
httpx
uvicorn
passlib
jinja2
aiofiles
python-multipart
pyjwt
bcrypt